+++ spec/features/users_can_mark_optional_tasks_as_not_applicable_spec.rb
@@ -4,11 +4,11 @@ RSpec.feature "Users can mark optional tasks as not applicable to a project" do
   let(:user_one) { create(:user, email: "user.one@education.gov.uk") }
 
   before do
-    mock_successful_api_responses(urn: 123456, ukprn: 123456)
+    mock_successful_api_responses(urn: 123456, ukprn: 12345678)
     sign_in_with_user(user_one)
   end
 
-  let(:project) { create(:project, urn: 123456, trust_ukprn: 123456) }
+  let(:project) { create(:project, urn: 123456, trust_ukprn: 12345678) }
   let!(:section) { create(:section, project: project) }
   let(:task) { create(:task, :not_applicable, title: "Not applicable task", section: section) }
   let!(:actions) { create_list(:action, 3, task: task, completed: true) }