+++ spec/models/project_spec.rb
@@ -71,6 +71,8 @@ RSpec.describe Project, type: :model do
     describe "#trust_ukprn" do
       it { is_expected.to validate_presence_of(:trust_ukprn) }
       it { is_expected.to validate_numericality_of(:trust_ukprn).only_integer }
+      it { is_expected.to allow_value(12345678).for(:trust_ukprn) }
+      it { is_expected.not_to allow_values(1234567, 123456789, 23456789).for(:trust_ukprn) }
 
       context "when no trust with that UKPRN exists in the API" do
         let(:no_trust_found_result) do