+++ CHANGELOG.md
@@ -55,6 +55,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Rename any references of "Delivery officer" to "Caseworker". Projects will
   lose any currently assigned delivery officers.
 - Add validation that checks URN is 6 digits long.
+- Add validation that checks UKPRN is 8 digits long and starts with a 1.
 
 ## [Release 1][release-1]